requirements.txt
# Core PyTorch dependencies
torch>=2.5.0
torchvision>=0.20.0
torchaudio>=2.5.0

# PyTorch Lightning
pytorch-lightning>=2.0.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0

# Logging and monitoring
tensorboard>=2.15.0
wandb>=0.16.0

# Configuration management
PyYAML>=6.0
hydra-core>=1.3.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: For advanced features
# albumentations>=1.3.0  # Uncomment for advanced image augmentations
# timm>=0.9.0           # Uncomment for additional model architectures
# transformers>=4.30.0   # Uncomment for additional tokenizer support
